Add information about which hostnames are behind Cloudflare
One of the main goals behind esnicheck.com is to encourage the adoption
of ESNI, so that major websites/providers support it, for the benefit of
all users as censors may be hesitant to block a technology that is used
by everyone and supported by major websites; censors have been hesitant
to block TCP/443 for this reason. However, given the state of ESNI right
now, most websites that do support it are behind Cloudflare.  While
Cloudflare's support for pushing privacy standards like DoH and ESNI
should be applauded, it is also worth nothing that this may be a
problem: Cloudflare should not be the only provider supporting ESNI as
for it to be successful, we need support from other providers as well.

This commit adds the text "Cloudflare" in front of the list of top
websites to indicate if they are behind Cloudflare (see the
is_cloudflare function in mostvisited/mostvisitedesni.py). The idea is
to understand and measure the current landscape of ESNI deployment on
the server side.