Improve the results page
This commit updates the results page to show more information about if
the host is behind Cloudflare. We are currently doing this for the
websites in mostvisited/ but now we do it for all queries.

This is an attempt to start addressing #1 to explain why it is a problem
if all websites that have ESNI enabled are running on Cloudflare.